change to uint64

+++ include/ilang/util/str_util.h
@@ -17,7 +17,7 @@ namespace ilang {
 /// (itoa is not part of the standard actually,
 /// so I hesitated whether to use it actually)
 /// on the other hand, snprintf only supports 8/10/16
-std::string IntToStrCustomBase(unsigned value, unsigned base, bool uppercase);
+std::string IntToStrCustomBase(uint64_t value, unsigned base, bool uppercase);
 
 /// Transform basic string to upper case.
 std::string StrToUpper(const std::string& str);
